requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.5.0
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
bcrypt==3.2.2
beanie==1.10.8
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==37.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-3-jet==1.0.8
django-adminlte-3==0.1.6
django-crispy-forms==2.0
django-filter==22.1
django-import-export==3.2.0
django-pandas==0.6.6
djangorestframework==3.14.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
et-xmlfile==1.1.0
ez-setup==0.9
fastapi==0.75.2
frozenlist==1.3.3
google-api-python-client==1.4.1
h11==0.13.0
httplib2==0.21.0
httptools==0.4.0
idna==3.3
Markdown==3.4.1
MarkupPy==1.14
motor==2.5.1
multidict==6.0.2
numpy==1.24.2
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==1.5.3
passlib==1.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.9.0
PyJWT==2.6.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.3
starlette==0.17.1
tablib==3.4.0
toml==0.10.2
twilio==8.0.0
typing_extensions==4.2.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
xlrd==2.0.1
XlsxWriter==3.1.0
xlwt==1.3.0
yarl==1.7.2
